+++ src/main/java/de/muspellheim/commons/fx/control/decoration/package-info.java
@@ -1,7 +1,0 @@
-/*
- * Muspellheim Commons FX
- * Copyright (c) 2019 Falko Schumann
- */
-
-/** Contains extension to ControlFX decoration. */
-package de.muspellheim.commons.fx.control.decoration;