+++ src/main/java/de/muspellheim/commons/fx/validation/Severity.java
@@ -3,5 +3,12 @@
  * Copyright (c) 2019 Falko Schumann
  */
 
-/** Contains form validation. */
 package de.muspellheim.commons.fx.validation;
+
+/** Severity of a message. */
+public enum Severity {
+  OK,
+  INFO,
+  WARNING,
+  ERROR
+}